Back Finalizado

+++ backend/controllers/tareasControllers.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/tareasModel')
 
 const getTareas = asyncHandler(async (req, res) => {
 
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user.id })
 
     res.status(200).json(tareas)
 })
@@ -15,7 +15,8 @@ const setTarea = asyncHandler(async (req, res) => {
     }
 
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto,
+        user: req.user.id
     })
     res.status(201).json(tarea)
 })
@@ -29,9 +30,14 @@ const updateTarea = asyncHandler(async (req, res) => {
         throw new Error('La tarea no fué encontrada')
     }
 
-    const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-
-    res.status(200).json(updatedTarea)
+    //verificamos que la tarea pertenezca al usuario logeado
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+    } else {
+        const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        res.status(200).json(updatedTarea)
+    }
 })
 
 const deleteTarea = asyncHandler(async (req, res) => {
@@ -43,10 +49,18 @@ const deleteTarea = asyncHandler(async (req, res) => {
         throw new Error('La tarea no fué encontrada')
     }
 
-    //const deletedTarea = await Tarea.findByIdAndDelete(req.params.id)
-    tarea.deleteOne()
+    //verificamos que la tarea pertenezca al usuario logeado
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+    } else {
+        tarea.deleteOne()
+        res.status(200).json({ id: tarea._id })
+    }
 
-    res.status(200).json({ id: tarea._id })
+    //const deletedTarea = await Tarea.findByIdAndDelete(req.params.id)
+    //tarea.deleteOne()
+    //res.status(200).json({ id: tarea._id })
 })
 
 module.exports = {
